requirements.txt
torch==1.10.2
torchvision==0.11.3
transformers==4.19.2
numpy==1.22.2
scikit_learn==1.1.1
Pillow==9.2.0
tqdm==4.62.3